change precision to 32

+++ radbert_pipe.py
@@ -33,7 +33,7 @@ def _load_vae(component_name, path, torch_dtype, variant):
 
 class FrozenCustomPipe:
     def __init__(self, use_freeze=True, path="/vol/ideadata/ce90tate/cxr_phrase_grounding/components",variant=None,
-                 llm_name="radbert", torch_dtype=torch.float16, device="cuda", save_attention=False, inpaint=False, accelerator=None):
+                 llm_name="radbert", torch_dtype=torch.float32, device="cuda", save_attention=False, inpaint=False, accelerator=None):
         self.device = device
         component_loader = {
             "text_encoder": _load_text_encoder,
